+++ pmd-plsql/src/test/java/net/sourceforge/pmd/lang/plsql/ast/PlsqlTreeDumpTest.java
@@ -51,4 +51,14 @@ class PlsqlTreeDumpTest extends BaseTreeDumpTest {
     void errorLoggingClause() {
         doTest("ErrorLoggingClause2779");
     }
+
+    @Test
+    void compoundTriggerWithAdditionalDeclarations() {
+        doTest("CompoundTriggerWithAdditionalDeclarations4270");
+    }
+
+    @Test
+    void exceptionHandlerTomKytesDespair() {
+        doTest("ExceptionHandlerTomKytesDespair");
+    }
 }